# Faster i256 parsing